requirements.txt
flask
openai-whisper
transformers
torch